CI: compile flows/ against a locally published build
Every flow pins the last release (org.virtuslab::orca:0.1.0), so compiling
the files as they sit on disk resolved that artifact from Maven Central: an
API change in the working tree could break all seven flows with the check
still green. CI never ran the check at all.

Each flow is now staged into a temp dir with its dep pin rewritten to the
version publishLocal just produced plus `//> using repository ivy2Local` —
the rewrite BuiltInFlows and `_seed_lib.sh --local` already apply — and
compiled from there. A `flow-scripts` CI job runs that suite alone with
ORCA_INTEGRATION=1.

Splits the old ScalaCliSmokeTest by what it needs: FlowScriptsCompileTest
(scala-cli only, so CI can run it) and ScalaCliSmokeTest (starts a real
flow, so it needs an authenticated claude and stays local-only). The latter
now runs in a seeded git repo — a flow refuses to start outside one, so it
had been failing.
