--- runner/src/test/scala/orca/runner/terminal/ScalaCliSmokeTest.scala
@@ -1,143 +1,0 @@
-package orca.runner.terminal
-import orca.testkit.TempDirs
-
-/** Publishes the library to the local Ivy cache, then exercises both the
-  * minimal smoke script and the real flow scripts under `flows/` via
-  * `scala-cli`. Gated on `ORCA_INTEGRATION` because it shells out to a real sbt
-  * + scala-cli and the fan-out across flows can take several minutes cold. CI
-  * sets the env var; local devs opt in when they want the check.
-  */
-class ScalaCliSmokeTest extends munit.FunSuite:
-
-  override def munitTests(): Seq[Test] =
-    if sys.env.contains("ORCA_INTEGRATION") then super.munitTests()
-    else Nil
-
-  override def munitTimeout: scala.concurrent.duration.Duration =
-    import scala.concurrent.duration.DurationInt
-    20.minutes
-
-  /** Memoises the publishLocal step across the suite (every script links
-    * against the same version). Runs once per suite, fails fast if sbt fails,
-    * and exposes the repo root plus the published version, read back from sbt
-    * because dynver derives it from git state rather than a static literal.
-    */
-  case class Published(repoRoot: os.Path, version: String)
-
-  private val publishedRepo = new munit.Fixture[Published]("publishedRepo"):
-    private var resolved: Published = null
-    override def apply(): Published = resolved
-    override def beforeAll(): Unit =
-      val repoRoot = findRepoRoot()
-      // `sbt --client` reuses a running sbt server when one exists, so the
-      // second invocation (print version) skips the JVM cold-start.
-      val publishResult = os
-        .proc("sbt", "--client", "publishLocal")
-        .call(cwd = repoRoot, check = false)
-      assertEquals(
-        publishResult.exitCode,
-        0,
-        s"publishLocal failed: ${publishResult.err.text()}"
-      )
-      val versionResult = os
-        .proc("sbt", "--client", "--error", "print version")
-        .call(cwd = repoRoot, check = false)
-      assertEquals(
-        versionResult.exitCode,
-        0,
-        s"reading version failed: ${versionResult.err.text()}"
-      )
-      // `print version` aggregates across every subproject; pick the first
-      // version line (starts with whitespace + a digit) — they're all the
-      // same dynver value. `sbt --client` decorates output with ANSI escapes,
-      // so strip them before matching.
-      val ansiEscape = "\u001b\\[[0-9;]*[A-Za-z]".r
-      val version = versionResult.out
-        .lines()
-        .iterator
-        .map(line => ansiEscape.replaceAllIn(line, "").trim)
-        .find(s => s.headOption.exists(_.isDigit))
-        .getOrElse(fail("could not parse orca version from sbt output"))
-      resolved = Published(repoRoot, version)
-
-  override def munitFixtures: Seq[munit.AnyFixture[?]] = Seq(publishedRepo)
-
-  // Keep in sync with V.scala (project/Dependencies.scala) — not reachable
-  // from test code (build-definition sources aren't on the runtime/test
-  // classpath), so pinned here as a literal. A stale value here fails loudly:
-  // the published library's TASTy won't parse under an older scala-cli pin.
-  private val scalaVersion = "3.8.4"
-
-  test(
-    "scala-cli runs a minimal script that links against the published library"
-  ):
-    val scriptDir = TempDirs.dir()
-    val script = scriptDir / "hello.sc"
-    val version = publishedRepo().version
-    os.write(
-      script,
-      s"""//> using scala $scalaVersion
-         |//> using repository ivy2Local
-         |//> using dep org.virtuslab::orca:$version
-         |//> using jvm 21
-         |
-         |import orca.{*, given}
-         |
-         |flow(args = OrcaArgs("smoke test")):
-         |  println(s"userPrompt=$$userPrompt")
-         |""".stripMargin
-    )
-
-    val runResult = os
-      .proc("scala-cli", "run", script.toString)
-      .call(cwd = scriptDir, check = false, mergeErrIntoOut = true)
-    val runOutput = runResult.out.text()
-    assertEquals(runResult.exitCode, 0, s"scala-cli run failed: $runOutput")
-    assert(
-      runOutput.contains("userPrompt=smoke test"),
-      s"expected stdout to contain 'userPrompt=smoke test', got: $runOutput"
-    )
-
-  /** Add new flow scripts here so they're picked up by the compile-check loop.
-    * Each entry is a path under `flows/` (the seed scripts copy these `.sc`
-    * files into the user's project at create-test-project time).
-    */
-  private val flowScripts: Seq[String] = Seq(
-    "implement.sc",
-    "implement-interactive.sc",
-    "implement-enhanced.sc",
-    "issue-pr-bugfix.sc",
-    "issue-pr.sc",
-    "epic.sc",
-    "simple.sc"
-  )
-
-  /** Each flow script is a real-world consumer of the public API, so a rename
-    * or signature change breaks them first. Compile-checking here closes the
-    * gap between sbt's internal compile (which doesn't see `flows/`) and what a
-    * fresh user runs.
-    */
-  for relPath <- flowScripts do
-    test(s"flows/$relPath compiles via scala-cli"):
-      val repoRoot = publishedRepo().repoRoot
-      val scriptPath = repoRoot / "flows" / relPath
-      val result = os
-        .proc("scala-cli", "compile", scriptPath.toString)
-        .call(cwd = repoRoot, check = false, mergeErrIntoOut = true)
-      assertEquals(
-        result.exitCode,
-        0,
-        s"scala-cli compile failed for flows/$relPath:\n${result.out.text()}"
-      )
-
-  /** Walk up from the test's working directory until we see a build.sbt. */
-  private def findRepoRoot(): os.Path =
-    @scala.annotation.tailrec
-    def loop(p: os.Path): os.Path =
-      if os.exists(p / "build.sbt") then p
-      else if p == p / os.up then
-        throw new RuntimeException(
-          "No build.sbt found walking up from " + os.pwd
-        )
-      else loop(p / os.up)
-    loop(os.pwd)
